requirements.txt
accelerate==1.12.0
annotated-doc==0.0.4
anyio==4.12.1
bitsandbytes==0.49.2
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
filelock==3.25.0
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.5.0
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
packaging
pillow==12.1.1
psutil==7.2.2
Pygments==2.19.2
PyYAML==6.0.3
regex==2026.2.28
rich==14.3.3
safetensors==0.7.0
sentencepiece==0.2.1
setuptools==80.10.2
shellingham==1.5.4
sympy==1.13.1
tokenizers==0.22.2
# torch==2.6.0+cu124
# torchaudio==2.6.0+cu124
# torchvision==0.21.0+cu124
tqdm==4.67.3
# transformers @ git+https://github.com/huggingface/transformers.git@5c1c72be5f864d10d0efe8ece0768d9ed6ee4fdd
transformers
llama-cpp-python
typer==0.24.1
typer-slim==0.24.0
typing_extensions==4.15.0
wheel==0.46.3
